+++ src/renderer/src/App.jsx
@@ -1,34 +1,15 @@
-import { useState } from 'react'
-import reactLogo from '../../../assets/react.svg'
-import './App.css'
+import React from "react";
+import TodoDate from "./components/TodoDate.jsx";
+import ItemList from "./components/ItemList.jsx";
+import { AppStateProvider } from "./AppContext.jsx";
 
 function App() {
-  const [count, setCount] = useState(0)
-
   return (
-    <div className="App">
-      <div>
-        <a href="https://vitejs.dev" target="_blank">
-          <img src="vite.svg" className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://reactjs.org" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </div>
-  )
+    <AppStateProvider>
+      <TodoDate />
+      <ItemList />
+    </AppStateProvider>
+  );
 }
 
-export default App
+export default App;
