+++ src/renderer/src/components/Progress.module.scss
